# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: wsh API
  version: "0.1.0"
  description: >
    wsh exposes terminal I/O via HTTP and WebSocket. This spec covers every
    endpoint, request/response schema, authentication, and error format.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Default local development server
security:
  - bearerAuth: []
  - queryAuth: []
...
paths:
...
  /input:
    post:
      operationId: postInput
      summary: Inject input bytes into the terminal PTY
      tags: [input]
      description: >
        Sends raw bytes to the terminal. The request body is forwarded
        verbatim to the PTY -- no JSON wrapping.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
          text/plain:
            schema:
              type: string
      responses:
        "204":
          description: Input accepted and forwarded to PTY.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          description: Failed to send input to PTY.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /screen:
    get:
      operationId: getScreen
      summary: Get current visible screen state
      tags: [terminal]
      parameters:
        - name: format
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Format"
      responses:
        "200":
          description: Current screen state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: Terminal parser unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scrollback:
    get:
      operationId: getScrollback
      summary: Get scrollback buffer contents
      tags: [terminal]
      parameters:
        - name: format
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Format"
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 100
      responses:
        "200":
          description: Scrollback buffer contents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrollbackResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: Terminal parser unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ws/raw:
    get:
      operationId: wsRaw
      summary: Raw binary WebSocket
      tags: [websocket]
      description: >
        Bidirectional byte stream mirroring the PTY. Output arrives as binary
        frames. Send binary or text frames to inject input.
      responses:
        "101":
          description: WebSocket upgrade successful.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /ws/json:
    get:
      operationId: wsJson
      summary: JSON request/response and event WebSocket
      tags: [websocket]
      description: >
        Unified request/response protocol with event subscriptions. After
        upgrade, server sends {"connected": true}. Client can immediately
        send method calls (get_screen, send_input, subscribe, etc.) and
        receive responses. Subscribing to events is optional. See
        websocket.md for the full protocol specification.
      responses:
        "101":
          description: WebSocket upgrade successful.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /overlay:
    post:
      operationId: createOverlay
      summary: Create an overlay
      tags: [overlay]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOverlayRequest"
      responses:
        "201":
          description: Overlay created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOverlayResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      operationId: listOverlays
      summary: List all overlays
      tags: [overlay]
      responses:
        "200":
          description: Array of all overlays.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Overlay"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: clearOverlays
      summary: Clear all overlays
      tags: [overlay]
      responses:
        "204":
          description: All overlays removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /overlay/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Overlay ID
    get:
      operationId: getOverlay
      summary: Get a single overlay
      tags: [overlay]
      responses:
        "200":
          description: The overlay.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Overlay"
        "404":
          description: Overlay not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      operationId: updateOverlay
      summary: Replace overlay spans
      tags: [overlay]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOverlayRequest"
      responses:
        "204":
          description: Overlay updated.
        "404":
          description: Overlay not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      operationId: patchOverlay
      summary: Move or reorder an overlay
      tags: [overlay]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchOverlayRequest"
      responses:
        "204":
          description: Overlay patched.
        "404":
          description: Overlay not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: deleteOverlay
      summary: Delete an overlay
      tags: [overlay]
      responses:
        "204":
          description: Overlay deleted.
        "404":
          description: Overlay not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /panel:
    post:
      operationId: createPanel
      summary: Create a panel
      tags: [panel]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePanelRequest"
      responses:
        "201":
          description: Panel created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePanelResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      operationId: listPanels
      summary: List all panels
      tags: [panel]
      responses:
        "200":
          description: Array of all panels sorted by position then z-order descending.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Panel"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: clearPanels
      summary: Clear all panels
      tags: [panel]
      responses:
        "204":
          description: All panels removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /panel/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Panel ID
    get:
      operationId: getPanel
      summary: Get a single panel
      tags: [panel]
      responses:
        "200":
          description: The panel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
        "404":
          description: Panel not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      operationId: updatePanel
      summary: Replace a panel
      tags: [panel]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePanelRequest"
      responses:
        "204":
          description: Panel replaced.
        "404":
          description: Panel not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      operationId: patchPanel
      summary: Partially update a panel
      tags: [panel]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchPanelRequest"
      responses:
        "204":
          description: Panel patched.
        "404":
          description: Panel not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: deletePanel
      summary: Delete a panel
      tags: [panel]
      responses:
        "204":
          description: Panel deleted.
        "404":
          description: Panel not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /quiesce:
    get:
      operationId: getQuiesce
      summary: Wait for terminal quiescence
      tags: [terminal]
      description: >
        Long-polls until the terminal has been idle for timeout_ms milliseconds,
        then returns a full screen state snapshot. Returns 408 if max_wait_ms
        is exceeded without quiescence.
      parameters:
        - name: timeout_ms
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
          description: Quiescence threshold in milliseconds
        - name: format
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Format"
        - name: max_wait_ms
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 30000
          description: Overall deadline in milliseconds
      responses:
        "200":
          description: Terminal is quiescent. Returns screen state snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuiesceResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "408":
          description: Deadline exceeded without quiescence.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /input/mode:
    get:
      operationId: getInputMode
      summary: Get current input mode
      tags: [input]
      responses:
        "200":
          description: Current input mode.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InputModeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /input/capture:
    post:
      operationId: postInputCapture
      summary: Switch to capture mode
      tags: [input]
      description: >
        In capture mode, keyboard input is broadcast to API subscribers
        but not forwarded to the PTY. Idempotent.
      responses:
        "204":
          description: Mode switched to capture.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /input/release:
    post:
      operationId: postInputRelease
      summary: Switch to passthrough mode
      tags: [input]
      description: >
        In passthrough mode, keyboard input goes to both API subscribers
        and the PTY. Idempotent.
      responses:
        "204":
          description: Mode switched to passthrough.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Authorization: Bearer <token>"
    queryAuth:
      type: apiKey
      in: query
      name: token
      description: "?token=<token> query parameter"
  responses:
    Unauthorized:
      description: No credentials provided.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Format:
      type: string
      enum: [plain, styled]
      default: styled
...
    ScreenResponse:
      type: object
      required: [epoch, first_line_index, total_lines, lines, cursor, cols, rows, alternate_active]
      properties:
        epoch: { type: integer, minimum: 0 }
        first_line_index: { type: integer, minimum: 0 }
        total_lines: { type: integer, minimum: 0 }
        lines: { type: array, items: { $ref: "#/components/schemas/FormattedLine" } }
        cursor: { $ref: "#/components/schemas/Cursor" }
        cols: { type: integer, minimum: 1 }
        rows: { type: integer, minimum: 1 }
        alternate_active: { type: boolean }
    ScrollbackResponse:
      type: object
      required: [epoch, lines, total_lines, offset]
      properties:
        epoch: { type: integer, minimum: 0 }
        lines: { type: array, items: { $ref: "#/components/schemas/FormattedLine" } }
        total_lines: { type: integer, minimum: 0 }
        offset: { type: integer, minimum: 0 }
...
    Overlay:
      type: object
      required: [id, x, y, z, spans]
      properties:
        id: { type: string }
        x: { type: integer, minimum: 0 }
        y: { type: integer, minimum: 0 }
        z: { type: integer }
        spans: { type: array, items: { $ref: "#/components/schemas/OverlaySpan" } }
    CreateOverlayRequest:
      type: object
      required: [x, y, spans]
      properties:
        x: { type: integer, minimum: 0 }
        y: { type: integer, minimum: 0 }
        z: { type: integer }
        spans: { type: array, items: { $ref: "#/components/schemas/OverlaySpan" } }
    CreateOverlayResponse:
      type: object
      required: [id]
      properties:
        id: { type: string }
    UpdateOverlayRequest:
      type: object
      required: [spans]
      properties:
        spans: { type: array, items: { $ref: "#/components/schemas/OverlaySpan" } }
    PatchOverlayRequest:
      type: object
      properties:
        x: { type: integer, minimum: 0 }
        y: { type: integer, minimum: 0 }
        z: { type: integer }

    # --- Panels ---
...
    CreatePanelRequest:
      type: object
      required: [position, height]
      properties:
        position: { $ref: "#/components/schemas/PanelPosition" }
        height: { type: integer, minimum: 1 }
        z: { type: integer }
        spans: { type: array, items: { $ref: "#/components/schemas/OverlaySpan" } }
    CreatePanelResponse:
      type: object
      required: [id]
      properties:
        id: { type: string }
...
    InputModeResponse:
      type: object
      required: [mode]
      properties:
        mode: { $ref: "#/components/schemas/InputMode" }

    # --- WebSocket Messages ---
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - auth_required
                - auth_invalid
                - not_found
                - overlay_not_found
                - panel_not_found
                - invalid_request
                - invalid_overlay
                - invalid_input_mode
                - invalid_format
                - unknown_method
                - channel_full
                - parser_unavailable
                - input_send_failed
                - quiesce_timeout
                - internal_error
            message: { type: string }